+++ NEWS.md
@@ -1,5 +1,8 @@
 # stringr (development version)
 
+* New `str_split_n()` function to extract only a single piece from a string
+  (#278, @bfgray3).
+
 # stringr 1.4.0
 
 * `str_interp()` now renders lists consistently independent on the presence of